Sections activity is made as Main Activity git status
