+++ app/src/main/AndroidManifest.xml
@@ -8,42 +8,50 @@
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
+
+
         <activity
-            android:name=".ChapterActivity"
+            android:name=".SectionsActivity"
             android:label="@string/app_name"
+            android:parentActivityName=".ChapterActivity"
             android:theme="@style/AppTheme.NoActionBar">
+
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
+
+
         </activity>
+
+        <activity
+            android:name=".ChapterActivity"
+            android:label="@string/chapter"
+            android:theme="@style/AppTheme.NoActionBar">
+            <meta-data
+                android:name="android.support.PARENT_ACTIVITY"
+                android:value=".SectionsActivity" />
+        </activity>
+
         <activity
             android:name=".AboutThirukkuralActivity"
             android:label="@string/about_thirukkural"
-            android:parentActivityName=".ChapterActivity"
+            android:parentActivityName=".SectionsActivity"
             android:theme="@style/AppTheme.NoActionBar">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.jpp.and_thirukkural.ChapterActivity" />
+                android:value="com.jpp.and_thirukkural.SectionsActivity" />
         </activity>
-        <activity
-            android:name=".SectionsActivity"
-            android:label="@string/book_index"
-            android:parentActivityName=".ChapterActivity"
-            android:theme="@style/AppTheme.NoActionBar">
-            <meta-data
-                android:name="android.support.PARENT_ACTIVITY"
-                android:value=".ChapterActivity" />
-        </activity>
+
         <activity
             android:name=".SearchResultsActivity"
             android:label="@string/search_results"
-            android:parentActivityName=".ChapterActivity"
+            android:parentActivityName=".SectionsActivity"
             android:theme="@style/AppTheme.NoActionBar">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.jpp.and_thirukkural.ChapterActivity" />
+                android:value="com.jpp.and_thirukkural.SectionsActivity" />
 
             <intent-filter>
                 <action android:name="android.intent.action.SEARCH" />
@@ -54,30 +62,26 @@
                 android:resource="@xml/searchable" />
         </activity>
 
-        <provider
-            android:name=".provider.ThirukkuralContentProvider"
-            android:authorities="com.jpp.and_thirukkural"
-            android:enabled="true"
-            android:exported="true" />
-
         <activity
             android:name=".FavoritesActivity"
             android:label="@string/fav_couplets"
-            android:parentActivityName=".ChapterActivity"
+            android:parentActivityName=".SectionsActivity"
             android:theme="@style/AppTheme">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
                 android:value="com.jpp.and_thirukkural.ChapterActivity" />
         </activity>
+
         <activity
             android:name=".SettingsActivity"
             android:label="@string/settings"
-            android:parentActivityName=".ChapterActivity"
+            android:parentActivityName=".SectionsActivity"
             android:theme="@style/AppTheme">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.jpp.and_thirukkural.ChapterActivity" />
+                android:value="com.jpp.and_thirukkural.SectionsActivity" />
         </activity>
+
         <activity
             android:name=".CoupletActivity"
             android:label="@string/couplet_detail"
@@ -87,6 +91,13 @@
                 android:name="android.support.PARENT_ACTIVITY"
                 android:value="com.jpp.and_thirukkural.ChapterActivity" />
         </activity>
+
+        <provider
+            android:name=".provider.ThirukkuralContentProvider"
+            android:authorities="com.jpp.and_thirukkural"
+            android:enabled="true"
+            android:exported="true" />
+
     </application>
 
 </manifest>
