+++ app/src/main/java/com/jpp/and_thirukkural/ChapterActivity.java
@@ -7,15 +7,11 @@ import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
-import android.support.design.widget.NavigationView;
 import android.support.design.widget.TabLayout;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
 import android.support.v4.app.FragmentPagerAdapter;
-import android.support.v4.view.GravityCompat;
 import android.support.v4.view.ViewPager;
-import android.support.v4.widget.DrawerLayout;
-import android.support.v7.app.ActionBarDrawerToggle;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.SearchView;
 import android.support.v7.widget.Toolbar;
@@ -38,8 +34,7 @@ import com.jpp.and_thirukkural.model.Section;
 
 import java.util.ArrayList;
 
-public class ChapterActivity extends AppCompatActivity
-        implements NavigationView.OnNavigationItemSelectedListener, SearchView.OnQueryTextListener {
+public class ChapterActivity extends AppCompatActivity implements SearchView.OnQueryTextListener {
 
     private static ArrayList<Chapter> allChapters;
     static DataLoadHelper dlh;
@@ -54,15 +49,9 @@ public class ChapterActivity extends AppCompatActivity
         setContentView(R.layout.activity_chapter);
         Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
+        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
 
-        DrawerLayout drawer = (DrawerLayout) findViewById(R.id.drawer_layout);
-        ActionBarDrawerToggle toggle = new ActionBarDrawerToggle(
-                this, drawer, toolbar, R.string.navigation_drawer_open, R.string.navigation_drawer_close);
-        drawer.setDrawerListener(toggle);
-        toggle.syncState();
-
-        NavigationView navigationView = (NavigationView) findViewById(R.id.nav_view);
-        navigationView.setNavigationItemSelectedListener(this);
+       //Load data
 
         dlh = new DataLoadHelper(getApplicationContext());
         allChapters = dlh.getAllChapters();
@@ -108,19 +97,9 @@ public class ChapterActivity extends AppCompatActivity
     }
 
     @Override
-    public void onBackPressed() {
-        DrawerLayout drawer = (DrawerLayout) findViewById(R.id.drawer_layout);
-        if (drawer.isDrawerOpen(GravityCompat.START)) {
-            drawer.closeDrawer(GravityCompat.START);
-        } else {
-            super.onBackPressed();
-        }
-    }
-
-    @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         // Inflate the menu; this adds items to the action bar if it is present.
-        getMenuInflater().inflate(R.menu.main, menu);
+        getMenuInflater().inflate(R.menu.menu_chapter, menu);
 
         // Associate searchable configuration with the SearchView
         SearchManager searchManager = (SearchManager) getSystemService(Context.SEARCH_SERVICE);
@@ -145,35 +124,6 @@ public class ChapterActivity extends AppCompatActivity
         }
 
         return super.onOptionsItemSelected(item);
-    }
-
-    @SuppressWarnings("StatementWithEmptyBody")
-    @Override
-    public boolean onNavigationItemSelected(MenuItem item) {
-        // Handle navigation view item clicks here.
-        int id = item.getItemId();
-
-        if (id == R.id.about_thirukkural) {
-            // Handle the camera action
-            Intent i = new Intent(this, AboutThirukkuralActivity.class);
-            startActivity(i);
-        } else if (id == R.id.book_index) {
-            Intent i = new Intent(this, SectionsActivity.class);
-            startActivity(i);
-        } else if (id == R.id.fav_couplets) {
-            Intent intent = new Intent(this, FavoritesActivity.class);
-            startActivity(intent);
-        } /*else if (id == R.id.about_app_creator) {
-            Intent intent = new Intent(this, AboutDeveloperActivity.class);
-            startActivity(intent);
-        } */else if (id == R.id.settings) {
-            Intent intent = new Intent(this, SettingsActivity.class);
-            startActivity(intent);
-        }
-
-        DrawerLayout drawer = (DrawerLayout) findViewById(R.id.drawer_layout);
-        drawer.closeDrawer(GravityCompat.START);
-        return true;
     }
 
     @Override
