+++ app/src/main/res/layout/content_section.xml
@@ -5,14 +5,14 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context=".ChapterActivity"
-    tools:showIn="@layout/app_bar_chapter">
+    tools:context=".SectionsActivity"
+    tools:showIn="@layout/app_bar_section">
+
 
     <android.support.v4.view.ViewPager
+        android:id="@+id/sectionsPager"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:id="@+id/chapterPager"
-        >
+        app:layout_behavior="@string/appbar_scrolling_view_behavior" />
 
-    </android.support.v4.view.ViewPager>
 </RelativeLayout>
