+++ app/src/main/res/layout/fragment_sections.xml
@@ -2,8 +2,7 @@
 xmlns:tools="http://schemas.android.com/tools"
 android:layout_width="match_parent"
 android:layout_height="match_parent"
-    tools:context="com.jpp.and_thirukkural.SectionsActivity$PlaceholderFragment"
-    android:paddingBottom="60dp">
+    tools:context="com.jpp.and_thirukkural.SectionsActivity$PlaceholderFragment">
 
 <ListView
     android:layout_width="fill_parent"
