+++ app/src/main/res/menu/menu_sections.xml
@@ -1,10 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.jpp.and_thirukkural.SectionsActivity">
+    xmlns:app="http://schemas.android.com/apk/res-auto">
     <item
-        android:id="@+id/action_settings"
-        android:orderInCategory="100"
-        android:title="@string/action_settings"
-        app:showAsAction="never" />
+        android:id="@+id/search_menu_item"
+        android:icon="@drawable/ic_menu_search"
+        app:showAsAction="collapseActionView|always"
+        android:title="@string/search"
+        app:actionViewClass="android.support.v7.widget.SearchView"/>
 </menu>
